LibWeb: Rename StyleRules.cpp -> StyleRule.cpp and tidy up
`append_with_to_string()` is no longer needed now that ComponentValue
has a Formatter.
