+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -50,7 +50,7 @@ set(SOURCES
     CSS/Parser/DeclarationOrAtRule.cpp
     CSS/Parser/Function.cpp
     CSS/Parser/Parser.cpp
-    CSS/Parser/StyleRules.cpp
+    CSS/Parser/StyleRule.cpp
     CSS/Parser/Token.cpp
     CSS/Parser/Tokenizer.cpp
     CSS/Percentage.cpp
